Make XmlDocumentationProvider public
